+++ src/calva_power_tools/extension.cljs
@@ -4,6 +4,7 @@
    [calva-power-tools.extension.life-cycle-helpers :as lc-helpers]
    [calva-power-tools.extension.when-contexts :as when-contexts]
    [calva-power-tools.tool.clay :as clay]
+   [calva-power-tools.tool.performance :as performance]
    [calva-power-tools.tool.snitch :as snitch]
    [calva-power-tools.tool.tools-deps :as deps]))
 
@@ -19,6 +20,7 @@
   (clay/activate!)
   (deps/activate!)
   (snitch/activate!)
+  (performance/activate!)
 
   (when-contexts/set-context!+ db/!app-db :calva-power-tools/active? true)